+++ src/main/java/name/divinityunbound/block/custom/UnholyGrassBlock.java
@@ -1,0 +1,64 @@
+package name.divinityunbound.block.custom;
+
+import com.mojang.serialization.MapCodec;
+import name.divinityunbound.block.ModBlockConstants;
+import name.divinityunbound.block.ModBlocks;
+import name.divinityunbound.block.entity.ModBlockEntities;
+import name.divinityunbound.block.entity.UnholyGrassBlockEntity;
+import name.divinityunbound.block.entity.UnholySilencerBlockEntity;
+import net.minecraft.block.*;
+import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.block.entity.BlockEntityTicker;
+import net.minecraft.block.entity.BlockEntityType;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemStack;
+import net.minecraft.screen.NamedScreenHandlerFactory;
+import net.minecraft.state.StateManager;
+import net.minecraft.state.property.BooleanProperty;
+import net.minecraft.state.property.Properties;
+import net.minecraft.util.ActionResult;
+import net.minecraft.util.Hand;
+import net.minecraft.util.ItemScatterer;
+import net.minecraft.util.hit.BlockHitResult;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.random.Random;
+import net.minecraft.world.World;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.Iterator;
+
+public class UnholyGrassBlock extends BlockWithEntity implements BlockEntityProvider {
+    protected final Random random = Random.create();
+    public static final BooleanProperty ENABLED = Properties.ENABLED;
+    public static final MapCodec<UnholyGrassBlock> CODEC = UnholyGrassBlock.createCodec(UnholyGrassBlock::new);
+    public UnholyGrassBlock(Settings settings) {
+        super(settings);
+        this.setDefaultState(this.stateManager.getDefaultState().with(ENABLED, false));
+    }
+    @Override
+    public BlockRenderType getRenderType(BlockState state) {
+        return BlockRenderType.MODEL;
+    }
+    @Override
+    protected MapCodec<? extends BlockWithEntity> getCodec() {
+        return CODEC;
+    }
+
+    @Override
+    protected void appendProperties(StateManager.Builder<Block, BlockState> builder) {
+        builder.add(ENABLED);
+    }
+    @Nullable
+    @Override
+    public BlockEntity createBlockEntity(BlockPos pos, BlockState state) {
+        return new UnholyGrassBlockEntity(pos, state);
+    }
+
+    @Nullable
+    @Override
+    public <T extends BlockEntity> BlockEntityTicker<T> getTicker(World world, BlockState state, BlockEntityType<T> type) {
+        return validateTicker(type, ModBlockEntities.UNHOLY_GRASS_BLOCK_ENTITY,
+                (world1, pos, state1, blockEntity) -> blockEntity.tick(world1, pos, state1));
+    }
+}